harmonic source

+++ process-data.sh
@@ -3,6 +3,11 @@
 echo "$(date) Job Selection" >> joblog
 ./gen-data.sh
 
+if [ $? -eq 1 ]; then
+	echo "Job selection cancellation received, stopping data processing."
+	exit 0;
+fi
+
 echo "$(date) Generating data plots." >> joblog
 rm plot-export/*
 ./gen-plot.py
